+++ Sources/MapboxNavigationCore/Map/Style/NavigationMapStyleManager.swift
@@ -1,5 +1,6 @@
 import _MapboxNavigationHelpers
 import Combine
+import CoreLocation
 import MapboxDirections
 @_spi(Experimental) import MapboxMaps
 import UIKit
@@ -246,7 +247,12 @@ final class NavigationMapStyleManager {
 
         guard let routes else { return }
 
-        updateRoutes(routes, config: config, featureProvider: routelineFeatureProvider)
+        updateRoutes(
+            routes,
+            routeProgress: routeProgress,
+            config: config,
+            featureProvider: routelineFeatureProvider
+        )
         updateWaypoints(
             routes: routes,
             legIndex: routeProgress?.legIndex,
@@ -300,11 +306,13 @@ final class NavigationMapStyleManager {
 
     private func updateRoutes(
         _ routes: NavigationRoutes,
+        routeProgress: RouteProgress?,
         config: MapStyleConfig,
         featureProvider: RouteLineFeatureProvider
     ) {
         let features = routeLineMapFeatures(
             routes: routes,
+            routeProgress: routeProgress,
             config: config,
             featureProvider: featureProvider,
             customizedLayerProvider: customizedLineLayerProvider,
@@ -663,15 +671,29 @@ final class NavigationMapStyleManager {
 
     private func routeLineMapFeatures(
         routes: NavigationRoutes,
+        routeProgress: RouteProgress? = nil,
         config: MapStyleConfig,
         featureProvider: RouteLineFeatureProvider,
         customizedLayerProvider: CustomizedTypeLayerProvider<LineLayer>,
         customPosition: LayerPosition?
     ) -> [(RouteLineStyleContent, MapFeature)] {
         var features: [(RouteLineStyleContent, MapFeature)] = []
+
+        let offset = if config.routeLineTracksTraversal,
+                        let routeProgress,
+                        !routeProgress.routeIsComplete,
+                        let coordinate = mapView.location.latestLocation?.coordinate,
+                        let shape = routes.mainRoute.route.shape,
+                        let distance = shape.distance(), distance > 0,
+                        let traversedDistance = shape.distance(to: coordinate)
+        {
+            traversedDistance / distance
+        } else {
+            0.0
+        }
         let mainRouteFeature = routes.mainRoute.route.routeLineMapFeatures(
             ids: .main,
-            offset: 0,
+            offset: offset,
             isSoftGradient: config.congestionConfiguration.displaySoftGradientForTraffic,
             isAlternative: false,
             config: config,